optional log normalizing of rows
mainly to work around buggy offsetting of row height for narrow diffs
causing larger diffs to force height to exceed maxHeight. not perfect
but better. plus, some data's just really scattered and logs can be more
intuitive. probably best to do it in data first but if only doing it for
fitting on screen it still makes sense here.